+++ CHANGES.md
@@ -1,6 +1,8 @@
 ## x.x.x
 
-* Add `resolve` endpoint to Namerd HTTP API 
+* Add `resolve` endpoint to Namerd HTTP API
+* Add `authority` metadata field to re-write HTTP host/:authority on demand
+* Add `setHost` parameter for Consul CatalogNamer to set `authority` metadata 
 * Add file-system based name interpreter.
 * Add auth `token` parameter to Consul Namer & Dtab Store